multi queue: Fix VMI creation in case non virtio present with multi vCPU
In case the following happen together:
1. Multi queue is enabled
2. There is more than one vCPU
3. There is at least one non virtio interface

the VMI would fail to be created with the following error:
"LibvirtError(Code=38, Domain=0, Message='Unable to create tap device tap0: Invalid argument')"

Multi queue is only implemented for virtio interfaces.
The tap device which is created, didn't take into account the
interface model when calculating the number of queues.
This caused the problem for non virtio interfaces.
Fix it by taking into account the interface model.

Signed-off-by: Or Shoval <oshoval@redhat.com>