conformance: Update imagePullPolicy
Since the used tag is devel and not latest,
the default imagePullPolicy used is IfNotPresent.
While developing, it means that in case the conformance image
was rebuilt and repushed, it won't get updated.

Changing to Always allows to overcome this problem.
In case the image already exists with the same hash, it won't
be downloaded again, so it even doesn't affect performance.

Signed-off-by: Or Shoval <oshoval@redhat.com>