instancetype: Apply PreferredMachineType during VM mutation
Previously the MachineType of a VirtualMachine was *always* populated by
the VirtualMachine mutation webhook, either by user input, cluster
configuration or a default. This would cause any attempt to later apply
the PreferredMachineType preference within the VirtualMachine controller
to fail as it was assumed this value was provided by the user.

This change corrects this by looking up any referenced preferences and
applying PreferredMachineType to the VirtualMachine before falling back
to the cluster config and default values. Any errors raised by this
lookup are ignored as they will be caught by the validation webhook
later during the admission.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>