Add download option to memory dump virtctl
The download option can get a memory dump to an existing
pvc or create a new one and get the memory dump to it.
It will wait for the memory dump to complete and then
will create an export crd for that pvc and download the memory dump
to a given output file path.

Signed-off-by: Shelly Kagan <skagan@redhat.com>