Introduce dvbuilder
Add the `dvbuilder` package in order to build DVs for functional tests.

`dvbuilder.NewDataVolume` is a manufacture for a `DataVolume`.

The `NewDataVolume` function receives and a list of `dvOptions`.

The current options are:
* `WithNamespace` to set the DV namespace
* `WithPVC` to add a PVCOption spec to the DataVolume
* `WithForceBindAnnotation` - to set the
"cdi.kubevirt.io/storage.bind.immediate.requested" annotation

There are more specific versions for the `WithSource` option:
* `WithRegistryURLSource` to add a DataVolumeSource to the DataVolume, with a registry and a URL
* `WithRegistryURLSourceAndPullMethod` to add a DataVolumeSource to the
DataVolume, with a registry and a URL + pull method
* `WithBlankImageSource` to add a blank DataVolumeSource to the DataVolume
* `WithPVCSource` to add a DataVolumeSource to the DataVolume, with a PVC source

```golang
   dv := NewDVBuilder(
           WithNamespace("myNamespace"),
           WithBlankImageSource(),
           WithPVC(storageClass, pvcSizeForRegistryImport, accessMode, volumeMode)
         )
```

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>