tests: ignore transient PSA warning events
Kubernetes internally uses a namespace informer
that might not be up to date after virt-controller relabeled the namespace
to use a 'privileged' policy, let's ignore those events in tests
as with enough time k8s will catch up and the pod will be created.

This patch has to be removed when KubeVirt will be able to run VMs
using the 'restricted' PSA level.

Signed-off-by: Antonio Cardace <acardace@redhat.com>