tests: Don't run HTTP tests against a non-HTTP server
The server that we use for testing the ability to connect to VMs
from the outside is a simple netcat process that prints the
string 'Hello World!' over and over again.

The version of curl in CentOS Stream 9 returns an error when
connecting to such a server:

  curl: (1) Received HTTP/0.9 when not allowed

We could add the --http0.9 option to the curl invocation, but
at the end of the day there's no point in using an HTTP client
to connect to a non-HTTP server when we already have a test
that uses netcat to verify connectivity.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>