fix stat virt-vnc and virt-serial0 return error when use debian:buster as the base image
the default base image for virt-handler is centos, which workdir is /.
but we use a custom debian:buster as the base image, which workdir is /root.

when we run `virtctl -n test vnc vm1`, it reports an error:
Can't access VMI vm1: Internal error occurred: dialing virt-handler: websocket: bad handshake
I digged into this problem, and found that the virt-vnc sock path in `getUnixSocketPath`
is `proc/3484989/root/var/run/kubevirt-private/a44829ef-d44c-4a64-bf8a-2bcf274bad12/virt-vnc`
without leading '/'.
when it runs `os.Stat` with this path, it will return an error "no such file or directory".

so, I think maybe we can add a leading '/' to make an absolute path,
it will work no matter what the workdir of the virt-handler image is.

Signed-off-by: huyinhou <huyinhou@bytedance.com>